Fix a health-clamp overflow and address review findings
`clamped_health_value` computed the raw total in `i32`, but `Health::current`
is a `u16` and `HealthDelta::delta` is a full-range `i32`, so the sum can
exceed `i32::MAX` by up to 65535. That panicked in debug builds and wrapped
to a bogus, possibly negative value in release builds — which the clamp then
silently accepted as 0. The arithmetic is now `i64`, which cannot overflow
for any input of those types, and `ClampedHealth::clamped_from` widens to
`Option<i64>` so the diagnostic still reports the true pre-clamp total. A
regression case covers `current: 90, delta: i32::MAX`.

Deduplication had no multi-entity coverage: the test helper hard-coded
entity 1. `deduped_outputs` now takes an entity with each weighted decision,
and `dedupe_separates_entities` pins that entity 1's two decisions collapse
while entity 2's single decision passes through, with exactly one
aggregation diagnostic naming entity 1. Single-entity callers go through a
`for_entity_one` helper, so the existing matrices are unchanged. The
property test stays single-entity; per-entity fan-out is pinned
deterministically rather than complicating its running-total oracle.

The frame-rollback property's action match listed three variants explicitly
and swallowed the rest with `_ => {}`. The remaining variants are now
listed, so a new `Action` fails to compile until the model makes a
deliberate decision about it.

Both `#[expect(clippy::cast_precision_loss)]` reasons claimed movement
counts are tiny and the conversion exact, which is not true of the `i64`
weights the properties generate. They now state the actual justification:
the converted weight only scales a vector that is renormalized immediately
afterwards, so lost precision cannot change the emitted direction. The
finding named the one in `to_decision`; the identical claim in `apply` is
corrected too rather than left standing.

`decide/tests.rs` reached the 400-line `module_max_lines` budget, so it is
split into `decide/tests/mod.rs` and `decide/tests/properties.rs`.

Docs: ADR-003's removal criteria now say the documented coverage gap
motivated the proptest amendment, criterion 1 reads "Sampling proves
insufficient", the extreme-values sentence is split in two, and the Option C
heading drops a stray "since". The DBSP guide's non-positive-weight bullet
now says zero-weight records are ignored as well as negative-weight ones,
matching the `weight <= 0` gate, and uses "handwritten".

+++ src/dbsp_sync/output/health.rs
@@ -116,7 +116,10 @@ pub(super) struct ClampedHealth {
     pub(super) value: u16,
     /// `Some(raw)` when the uncapped total fell outside the valid range and
     /// was clamped; `None` when no clamping was needed.
-    pub(super) clamped_from: Option<i32>,
+    ///
+    /// Widened to `i64` because the raw total does not fit in `i32`: a
+    /// `u16` current health plus an `i32::MAX` delta exceeds `i32::MAX`.
+    pub(super) clamped_from: Option<i64>,
 }
 
 /// Computes the clamped `Health::current` for a delta, returning `None` if the
@@ -124,10 +127,16 @@ pub(super) struct ClampedHealth {
 ///
 /// Pure: it neither logs nor mutates. The caller reports any clamping using the
 /// returned [`ClampedHealth::clamped_from`].
+///
+/// The raw total is computed in `i64`. `Health::current` is a `u16` and
+/// `HealthDelta::delta` is a full-range `i32`, so their sum can exceed
+/// `i32::MAX` — computing it in `i32` panics in debug builds and wraps to a
+/// bogus (possibly negative) value in release builds, which the clamp would
+/// then silently accept. `i64` cannot overflow for any input of those types.
 pub(super) fn clamped_health_value(delta: &HealthDelta, health: &Health) -> Option<ClampedHealth> {
-    let current = i32::from(health.current);
-    let max = i32::from(health.max);
-    let raw = current + delta.delta;
+    let current = i64::from(health.current);
+    let max = i64::from(health.max);
+    let raw = current + i64::from(delta.delta);
     let new_value = raw.clamp(0, max);
     let clamped_from = (raw != new_value).then_some(raw);
     // `new_value` is `raw` clamped to `0..=health.max`, and `health.max` is a
